Applications structures are converted to use 64-bit integers

+++ pkg/monitor/applications/pmdb.go
@@ -30,22 +30,22 @@ type RaftInfo struct {
 	State                    string    `json:"state"`
 	FollowerReason           string    `json:"follower-reason"`
 	ClientRequests           string    `json:"client-requests"`
-	Term                     int       `json:"term" type:"gauge" metric:"PMDB_term"`
-	CommitIdx                int       `json:"commit-idx" type:"gauge" metric:"PMDB_commitIdx"`
-	LastApplied              int       `json:"last-applied" type:"gauge" metric:"PMDB_last_applied"`
-	LastAppliedCumulativeCrc int       `json:"last-applied-cumulative-crc" type:"gauge" metric:"PMDB_last_applied_cumulative_crc"`
-	NewestEntryIdx           int       `json:"newest-entry-idx" type:"gauge" metric:"PMDB_newest_entry_idx"`
-	NewestEntryTerm          int       `json:"newest-entry-term" type:"gauge" metric:"PMDB_newest_entry_term"`
-	NewestEntryDataSize      int       `json:"newest-entry-data-size" type:"gauge" metric:"PMDB_newest_entry_data_size"`
-	NewestEntryCrc           int       `json:"newest-entry-crc" type:"gauge" metric:"PMDB_newest_entry_crc"`
+	Term                     int64     `json:"term" type:"gauge" metric:"PMDB_term"`
+	CommitIdx                int64     `json:"commit-idx" type:"gauge" metric:"PMDB_commitIdx"`
+	LastApplied              int64     `json:"last-applied" type:"gauge" metric:"PMDB_last_applied"`
+	LastAppliedCumulativeCrc int64     `json:"last-applied-cumulative-crc" type:"gauge" metric:"PMDB_last_applied_cumulative_crc"`
+	NewestEntryIdx           int64     `json:"newest-entry-idx" type:"gauge" metric:"PMDB_newest_entry_idx"`
+	NewestEntryTerm          int64     `json:"newest-entry-term" type:"gauge" metric:"PMDB_newest_entry_term"`
+	NewestEntryDataSize      int64     `json:"newest-entry-data-size" type:"gauge" metric:"PMDB_newest_entry_data_size"`
+	NewestEntryCrc           uint64    `json:"newest-entry-crc" type:"gauge" metric:"PMDB_newest_entry_crc"`
 	DevReadLatencyUsec       Histogram `json:"dev-read-latency-usec" type:"histogram" metric:"dev_read_latency_usec"`
 	DevWriteLatencyUsec      Histogram `json:"dev-write-latency-usec" type:"histogram" metric:"dev_write_latency_usec"`
 	FollowerStats            []struct {
 		PeerUUID    string `json:"peer-uuid"`
 		LastAckMs   int    `json:"ms-since-last-ack"`
 		LastAck     Time   `json:"last-ack"`
-		NextIdx     int    `json:"next-idx"`
-		PrevIdxTerm int    `json:"prev-idx-term"`
+		NextIdx     int64  `json:"next-idx"`
+		PrevIdxTerm int64  `json:"prev-idx-term"`
 	} `json:"follower-stats,omitempty"`
 	CommitLatencyMsec Histogram `json:"commit-latency-msec"`
 	ReadLatencyMsec   Histogram `json:"read-latency-msec"`
